updated module for separation of functions for upload/download speeds
i think importing the functions into the client and server modules is needed in order to display the network stats. also, i know it only says upload and download, but there is also the file transfer and file size metrics that are displayed, which i think satisfies our 3 metric need.